refactor: change path for Include in .ssh/config
The generated "Include ..." path inside .ssh/config is now more simple,
now the include is a wildcard applied for the synced folder instead the
paths for all single files.

BEFORE

Include <my-folder>/<my-file-00>
Include <my-folder>/<my-file-01>
...

AFTER
Include <my-folder>/*

In additional the logic of the sync command is splitted inside controllers package.
